Add files via upload
In Neyman-Pearson statistical tests, the probability of statistical errors depends on the
number of observations. In other words, by changing the number of observations, the
probability of statistical errors can also be altered. However, in such tests, we keep the
number of observations constant and reduce the probability of statistical errors.
Neyman-Pearson tests, especially Uniform Most Powerful (UMP) tests, are obtained
when we can change the probability of errors, and the sample size is fixed. But in
sequential tests, some methods are used to optimize sample size based on fixed
probability of type I and type II errors. These tests are based on the likelihood ratio tests.